terraform/function/pyproject.toml
[tool.poetry]
name = "reddit-top-commenter"
version = "0.1.0"
description = ""
authors = ["MichaelDickinson20I <mdickinson@twentyideas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
pytz = "^2024.1"
praw = "^7.7.1"
openai = "^1.42.0"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
